update chkdsk to version 0.9.2
git-svn-id: http://svn.code.sf.net/p/fdos/code/trunk@113 d9247ac3-808d-4053-b241-f1a6b6229609

+++ ENGINE/Cache/phscache.h
@@ -1,12 +1,12 @@
 #ifndef PHYSICAL_CACHE_H_
 #define PHYSICAL_CACHE_H_
 
-unsigned long InitialisePhysicalCache(void);
+unsigned InitialisePhysicalCache(void);
 void ClosePhysicalCache(void);
-int GetPhysicalMemType(unsigned long physblock);
+int GetPhysicalMemType(unsigned physblock);
 BOOL IsEMSCached(void);
-BOOL RemapXMSBlock(int logblock, unsigned long physblock);
-BOOL RemapEMSBlock(int logblock, unsigned long physblock);
+BOOL RemapXMSBlock(int logblock, unsigned physblock);
+BOOL RemapEMSBlock(int logblock, unsigned physblock);
 
 #define CACHEBLOCKSIZE EMS_PAGE_SIZE
